Revert "Feature/create events"
