--- lib/InitialShiftsInfo.js
@@ -1,23 +1,0 @@
-const initialShiftsInfo = [
-  {
-    id: 0,
-    shiftName: "Morning Shift 🌅",
-    start: "00:00",
-    end: "00:00",
-  },
-  {
-    id: 1,
-    shiftName: "Afternoon shift 🌞",
-    start: "00:00",
-    end: "00:00",
-  },
-
-  {
-    id: 3,
-    shiftName: "Night Shift 🌙",
-    start: "00:00",
-    end: "00:00",
-  },
-];
-
-export default initialShiftsInfo;
