+++ package.json
@@ -15,7 +15,6 @@
     "date-fns-tz": "^2.0.0",
     "eslint": "8.35.0",
     "eslint-config-next": "13.2.3",
-    "immer": "^10.0.2",
     "next": "13.2.3",
     "react": "18.2.0",
     "react-big-calendar": "^1.8.1",
@@ -23,7 +22,6 @@
     "react-dom": "18.2.0",
     "react-hook-form": "^7.45.0",
     "styled-components": "^5.3.8",
-    "use-local-storage-state": "^18.3.3",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
