+++ pages/index.js
@@ -1,18 +1,14 @@
 import Heading from "../components/Heading";
 // import SubHeading from "../components/SubHeading";
 import Calendar from "../components/Calendar";
+import events from "../lib/db";
 
-export default function Home({ events, setEvents, shiftsInfo, setShiftsInfo }) {
+export default function Home() {
   return (
     <main>
       <Heading>My Calendar</Heading>
       {/* <SubHeading></SubHeading> */}
-      <Calendar
-        events={events}
-        shiftsInfo={shiftsInfo}
-        setShiftsInfo={setShiftsInfo}
-        setEvents={setEvents}
-      />
+      <Calendar events={events} />
     </main>
   );
 }
